Fix descriptions of projects for maven central publishing
The descriptions of the projects were eagerly evaluated, before they
were set by the build.gradle in each project.
Putting them in a groovy string will lazily-evaluate them at the point
when they are accessed.

Also add a the checkMavenCentralRequirements task to the check lifecycle
task, so the validity is always being checked before a merge to main.