[ACC-1561] Ensure that version of Fabric8 kubernetes client matches the spring cloud dependency
With a too recent kubernetes client version, using it with spring cloud results in a runtime error because the fabric8 API has changed in an incompatible way.

To fix this alignment issue, we add an internal dependency on spring-cloud-dependencies,
but make sure not to expose that platform publicly.
If the platform is exposed to consumers, that will lead to all
dependencies specified in the platform to be set/upgraded to that version. We
don't use most of the specified dependencies, so we should not pin them
down for consumers.

```
***************************
APPLICATION FAILED TO START
***************************
Description:
An attempt was made to call a method that does not exist. The attempt was made from the following location:
    org.springframework.cloud.kubernetes.fabric8.Fabric8AutoConfiguration.kubernetesClientConfig(Fabric8AutoConfiguration.java:65)
The following method did not exist:
    'io.fabric8.kubernetes.client.ConfigFluent io.fabric8.kubernetes.client.ConfigBuilder.withMasterUrl(java.lang.String)'
The calling method's class, org.springframework.cloud.kubernetes.fabric8.Fabric8AutoConfiguration, was loaded from the following location:
    jar:file:/workspace/BOOT-INF/lib/spring-cloud-kubernetes-fabric8-autoconfig-3.1.3.jar!/org/springframework/cloud/kubernetes/fabric8/Fabric8AutoConfiguration.class
The called method's class, io.fabric8.kubernetes.client.ConfigBuilder, is available from the following locations:
    jar:file:/workspace/BOOT-INF/lib/kubernetes-client-api-6.13.3.jar!/io/fabric8/kubernetes/client/ConfigBuilder.class
The called method's class hierarchy was loaded from the following locations:
    io.fabric8.kubernetes.client.ConfigBuilder: file:/workspace/BOOT-INF/lib/kubernetes-client-api-6.13.3.jar
    io.fabric8.kubernetes.client.ConfigFluent: file:/workspace/BOOT-INF/lib/kubernetes-client-api-6.13.3.jar
    io.fabric8.kubernetes.client.SundrioConfigFluent: file:/workspace/BOOT-INF/lib/kubernetes-client-api-6.13.3.jar
    io.fabric8.kubernetes.api.builder.BaseFluent: file:/workspace/BOOT-INF/lib/kubernetes-model-common-6.13.3.jar
Action:
Correct the classpath of your application so that it contains compatible versions of the classes org.springframework.cloud.kubernetes.fabric8.Fabric8AutoConfiguration and io.fabric8.kubernetes.client.ConfigBuilder
```